% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/basic_functions.R
\name{create_direct}
\alias{create_direct}
\title{produce image directory}
\usage{
create_direct()
}
\description{
This function calls the image directories.
}
\examples{
create_proj()
}
\keyword{create}
