% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/basic_functions.R
\name{create_proj}
\alias{create_proj}
\title{creating a workspace}
\usage{
create_proj()
}
\description{
This function builds a simply project scope for image processing work.
}
\examples{
create_proj()
}
\keyword{create}
