% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/basic_functions.R
\name{download_vid}
\alias{download_vid}
\title{download video from link}
\usage{
download_vid(what, home, sbj)
}
\arguments{
\item{what}{the link to the video.}

\item{home}{the directory it will go to.}

\item{sbj}{a key work for the video to be named.}
}
\description{
This function pulls a video from the web to work with; you must run create_proj first if you omit sbj.
}
\keyword{download}
