% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/basic_functions.R
\name{etl_frames}
\alias{etl_frames}
\title{extract frames and produce an image bank from a video file}
\usage{
etl_frames(vid, height, width, chnl, output)
}
\arguments{
\item{vid}{the link to the video.}

\item{height}{the directory it will go to.}

\item{width}{a key work for the video to be named.}

\item{chnl}{a key work for the video to be named.}

\item{output}{a key work for the video to be named.}
}
\description{
This function extracts frames from video and places them in your workspace with specific parameters.
}
\examples{

}
\keyword{etl}
